Update agg examples
