+++ java-client/flight-examples/src/main/java/io/deephaven/client/examples/AggregateAllExample.java
@@ -1,30 +1,18 @@
 package io.deephaven.client.examples;
 
 import io.deephaven.api.agg.spec.AggSpec;
-import io.deephaven.client.impl.FlightSession;
-import io.deephaven.client.impl.TableHandle;
-import io.deephaven.qst.LabeledValue;
-import io.deephaven.qst.LabeledValues;
-import io.deephaven.qst.column.header.ColumnHeader;
-import io.deephaven.qst.column.header.ColumnHeaders8;
-import io.deephaven.qst.table.InMemoryKeyBackedInputTable;
 import io.deephaven.qst.table.LabeledTables;
 import io.deephaven.qst.table.LabeledTables.Builder;
-import io.deephaven.qst.table.TableHeader;
 import io.deephaven.qst.table.TableSpec;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
-import picocli.CommandLine.Option;
 
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import java.util.Random;
-import java.util.concurrent.TimeUnit;
 
 @Command(name = "aggregate-all", mixinStandardHelpOptions = true,
         description = "Aggregate all examples", version = "0.1.0")
-class AggregateAllExample extends FlightExampleBase {
+class AggregateAllExample extends AggInputBase {
 
     private static List<AggSpec> aggSpecs() {
         return Arrays.asList(
@@ -55,43 +43,9 @@ class AggregateAllExample extends FlightExampleBase {
         );
     }
 
-    @Option(names = {"--input-table-size"}, description = "The input table size, default ${DEFAULT-VALUE}",
-            defaultValue = "1000")
-    int inputTableSize;
-
-    @Option(names = {"--num-groups"}, description = "The number of groups, default ${DEFAULT-VALUE}",
-            defaultValue = "10")
-    int numGroups;
-
-    @Option(names = {"--update-percentage"},
-            description = "The update percentage per input table row per cycle, default ${DEFAULT-VALUE}",
-            defaultValue = "0.01")
-    double updatePercentage;
-
-    @Option(names = {"--sleep-millis"}, description = "The sleep milliseconds between cycles, default ${DEFAULT-VALUE}",
-            defaultValue = "100")
-    long sleepMillis;
-
     @Override
-    protected void execute(FlightSession flight) throws Exception {
-        final ColumnHeader<Integer> inputKey = ColumnHeader.ofInt("InputKey");
-        final ColumnHeader<Integer> groupKey = ColumnHeader.ofInt("GroupKey");
-        final ColumnHeader<Byte> u = ColumnHeader.ofByte("U");
-        final ColumnHeader<Short> v = ColumnHeader.ofShort("V");
-        final ColumnHeader<Integer> w = ColumnHeader.ofInt("W");
-        final ColumnHeader<Long> x = ColumnHeader.ofLong("X");
-        final ColumnHeader<Float> y = ColumnHeader.ofFloat("Y");
-        final ColumnHeader<Double> z = ColumnHeader.ofDouble("Z");
-
+    public LabeledTables labeledTables(TableSpec base) {
         final Builder builder = LabeledTables.builder();
-
-        final TableSpec spec = InMemoryKeyBackedInputTable.of(TableHeader.of(inputKey, groupKey, u, v, w, x, y, z),
-                Collections.singletonList(inputKey.name()));
-        builder.putMap("base", spec);
-        builder.putMap("countBy", spec.countBy("Count", groupKey.name()).sort(groupKey.name()));
-        builder.putMap("selectDistinct", spec.selectDistinct(groupKey.name()).sort(groupKey.name()));
-
-        // todo: countBy
         for (AggSpec aggSpec : aggSpecs()) {
             final String name = aggSpec.toString()
                     .replace('=', '_')
@@ -99,50 +53,10 @@ class AggregateAllExample extends FlightExampleBase {
                     .replace('}', '_')
                     .replace(' ', '_')
                     .replace('.', '_');
-            final TableSpec tableSpec = spec.aggAllBy(aggSpec, groupKey.name()).sort(groupKey.name());
+            final TableSpec tableSpec = base.aggAllBy(aggSpec, GROUP_KEY.name()).sort(GROUP_KEY.name());
             builder.putMap(name, tableSpec);
         }
-        final LabeledTables tables = builder.build();
-
-        final LabeledValues<TableHandle> results = flight.session().batch().execute(tables);
-        try {
-            for (LabeledValue<TableHandle> result : results) {
-                flight.session().publish(result.name(), result.value()).get(5, TimeUnit.SECONDS);
-            }
-
-            final ColumnHeaders8<Integer, Integer, Byte, Short, Integer, Long, Float, Double> headers =
-                    inputKey.header(groupKey).header(u).header(v).header(w).header(x).header(y).header(z);
-
-            final TableHandle base = results.get("base");
-
-            final Random random = new Random();
-
-            final int sizeGuess = (int) Math.round(inputTableSize * updatePercentage);
-
-            while (true) {
-                final ColumnHeaders8<Integer, Integer, Byte, Short, Integer, Long, Float, Double>.Rows rows =
-                        headers.start(sizeGuess);
-                for (int i = 0; i < inputTableSize; ++i) {
-                    if (random.nextDouble() > updatePercentage) {
-                        continue;
-                    }
-                    final int group = random.nextInt(numGroups);
-                    final byte u_ = (byte) random.nextInt();
-                    final short v_ = (short) random.nextInt();
-                    final int w_ = random.nextInt();
-                    final long x_ = random.nextLong();
-                    final float y_ = random.nextFloat();
-                    final double z_ = random.nextDouble();
-                    rows.row(i, group, u_, v_, w_, x_, y_, z_);
-                }
-                flight.addToInputTable(base, rows.newTable(), bufferAllocator);
-                Thread.sleep(sleepMillis);
-            }
-        } finally {
-            for (LabeledValue<TableHandle> result : results) {
-                result.value().close();
-            }
-        }
+        return builder.build();
     }
 
     public static void main(String[] args) {
